requirements.txt
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
certifi==2026.2.25
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
openpyxl==3.1.5
pydantic==2.12.5
pydantic_core==2.41.5
python-docx==1.2.0
python-dotenv==1.2.2
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
